requirements.txt
requests==2.31.0
python-dotenv==1.0.1
agno==2.2.6
openai==2.6.1
